% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rungo.R
\name{arraygo}
\alias{arraygo}
\title{Gene Ontology}
\usage{
arraygo(contractid, sampledata, workdir = NULL, myrds, contrastm)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}

\item{myrds}{List. An RDS file with lists of DM results}

\item{contrastsm}{Matrix. A matrix of contrasts}
}
\value{
CSV spreadsheets with Gene Ontology results.
}
\description{
A function to run gene ontology analysis from DM results
}
