% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundm.R
\name{arraymodel}
\alias{arraymodel}
\title{RunDM}
\usage{
arraymodel(contractid, mdata, sampledata, contrastl, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{mdata}{Data Frame. Cleaned, quality controlled M-values}

\item{sampledata}{Data Frame. The project samplesheet. Generate manually or in arraydm::readdata().}

\item{workdir}{Character. Where to save the results (Default: working directory)}

\item{contrasts}{List. Named list with contrasts to apply}
}
\value{
A data frame of modelling results
}
\description{
A function to run differential methylation
}
