% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraynorm.R
\name{arraynorm}
\alias{arraynorm}
\title{Normalises methylation array data}
\usage{
arraynorm(contractid, arraydata, mdataraw, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{The name of the contract}

\item{arraydata}{An RGChannelSet generated manually, or in arraydm::readdata()}

\item{mdataraw}{A filtered data frame of M-values}

\item{sampledata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata()}

\item{workdir}{Where to save the plots. (Default: working directory)}
}
\value{
PreQC plots for detection p-value, quality filtered probe set
}
\description{
Normalises methylation array data
}
