% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpaths.R
\name{arraypaths}
\alias{arraypaths}
\title{Pathways Analysis}
\usage{
arraypaths(contractid, sampledata, workdir = NULL, myrds, contrastm)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}

\item{myrds}{List. An RDS file with a list of DM results}

\item{contrastsm}{Matrix. A matrix of contrasts}
}
\value{
CSV spreadsheets with Reactome pathways results and enrichment results. Dot plot and barplots of enrichment results.
}
\description{
A function to run pathways analysis from DM results
}
