% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayqc.R
\name{arraypostqc}
\alias{arraypostqc}
\title{A function to qc array data2}
\usage{
arraypostqc(contractid, swandata, pdata, sampledata, workdir = NULL, outprefix)
}
\arguments{
\item{contractid}{The name of the contract}

\item{swandata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata()}

\item{pdata}{Detection P-values.}

\item{sampledata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata()}

\item{workdir}{Path to output location. (Default is working directory)}
}
\value{
Testing return2
}
\description{
A function to qc array data2
}
