% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayqc.R
\name{arraypreqc}
\alias{arraypreqc}
\title{A function to qc array data}
\usage{
arraypreqc(contractid, arraydata, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{The name of the contract}

\item{arraydata}{An RGChannelSet generated manually, or in arraydm::readdata()}

\item{sampledata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata()}

\item{workdir}{Path to output location. (Default is working directory)}
}
\value{
If PreQC is chosen, plots for detection p-value and a quality filtered probe set. For the postQC configuration function returns M and B values, the RDS file and postQC plots.
}
\description{
A function to qc array data
}
