% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{arrayread}
\alias{arrayread}
\title{Read Data}
\usage{
arrayread(contractid, workdir, samplesheet)
}
\arguments{
\item{contractid}{Character. The name of the contract}

\item{workdir}{Character. The path of the working directory}

\item{samplesheet}{Character. The suffix of the sample sheet}
}
\value{
The readdata object containing the methylation experiment and accompanying. Global objects 'epic' and 'targets' with the formatted inputs.
}
\description{
A function to load array data
}
