% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdm.R
\name{arraysumm}
\alias{arraysumm}
\title{Summarise Models}
\usage{
arraysumm(contractid, bdata, sampledata, workdir = NULL, contrastm)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{bdata}{Data Frame. Cleaned, quality controlled B-values}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}

\item{contrastm}{Matrix. A matrix of contrasts}

\item{contrasts}{List. A named list with contrasts to apply}
}
\value{
Annotated data frames of DM results, heatmaps and volcano plots
}
\description{
A function to collate and report differential methylation results
}
