% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargenes.R
\name{arrayvar}
\alias{arrayvar}
\title{A function to run vatiance analysis}
\usage{
arrayvar(contractid, mdata, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{mdata}{Data Frame. A filtered data frame of M-values}

\item{sampledata}{Data Frame. A data frame with the samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}
}
\value{
Outputs heatmaps and variance spreadsheets to your WD
}
\description{
A function to run vatiance analysis
}
