% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraydm.R
\name{auto_arraydm}
\alias{auto_arraydm}
\title{A wrapper to run all functions in sequence}
\usage{
auto_arraydm(contractid, sampledata, workdir, contrastlist)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{sampledata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Where to save the plots. (Default: working directory)}

\item{contrastlist}{List. A named list of comparisons to undertake.}
}
\value{
Full suite of differential methylation analysis
}
\description{
A wrapper to run all functions in sequence
}
