% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnorm.R
\name{.plotNorm}
\alias{.plotNorm}
\title{Array Normalisation}
\usage{
.plotNorm(contractid, rawdata, quantdata, swandata, sampledata, workdir)
}
\arguments{
\item{rawdata}{Data Frame. Raw M-values}

\item{quantdata}{Quantile normalised array.}

\item{swandata}{SWAN normalised array.}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata()}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}

\item{contract}{The name of the contract or similar ID}
}
\value{
Default SWAN Normalised array and data distribution
}
\description{
Plots normalised data. Best applied after arraydm::arraypreqc() where most of the inputs are generated.
}
