% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvolcano.R
\name{.plotVolc}
\alias{.plotVolc}
\title{A function to generate volcano plots from DM results}
\usage{
.plotVolc(contractid, sampledata, workdir = NULL, contrastm)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata().}

\item{workdir}{Charater. Where to save the plots. (Default: working directory)}

\item{contrastm}{Matrix. A matrix of contrasts}
}
\value{
One volcano plot per comparison
}
\description{
A function to generate volcano plots from DM results
}
