% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{A function to load array data}
\usage{
readdata(contractid, workdir, samplesheet)
}
\arguments{
\item{contractid}{The name of the contract}

\item{workdir}{The path of the working directory}

\item{samplesheet}{The suffix of the sample sheet}
}
\value{
The readdata object containing the methylation experiment and accompanying annotations
}
\description{
A function to load array data
}
