% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraynorm.R
\name{arraynorm}
\alias{arraynorm}
\title{Normalises methylation array data}
\usage{
arraynorm(contractid, arraydata, mdataraw, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract}

\item{arraydata}{An RGChannelSet generated manually, or in arraydm::readdata()}

\item{mdataraw}{Data Frame. A filtered data frame of M-values}

\item{sampledata}{Data Frame. The project samplesheet, generate manually or in arraydm::readdata()}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}
}
\value{
PreQC plots for detection p-value, quality filtered probe set. 'swannorm' global object.
}
\description{
Normalises methylation array data
}
