% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayqc.R
\name{arraypostqc}
\alias{arraypostqc}
\title{A function to qc array data2}
\usage{
arraypostqc(contractid, swandata, pdata, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract}

\item{swandata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata()}

\item{pdata}{Data Frame. Detection P-values.}

\item{sampledata}{Data Frame. The project samplesheet, generate manually or in arraydm::readdata()}

\item{workdir}{Character. Path to output location. (Default is working directory)}
}
\value{
M and B values, the RDS file and postQC plots.
}
\description{
A function to qc array data2
}
