% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayqc.R
\name{arraypreqc}
\alias{arraypreqc}
\title{Array QC}
\usage{
arraypreqc(contractid, arraydata, sampledata, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract}

\item{arraydata}{An RGChannelSet generated manually, or in arraydm::readdata()}

\item{sampledata}{Data Frame. The project samplesheet, generate manually or in arraydm::readdata()}

\item{workdir}{Character. Path to output location. (Default is working directory)}
}
\value{
Plots for detection p-value and a quality filtered probe set.
}
\description{
Functions to qc array data and summarise qc results.
}
