% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumiqc.R
\name{.plotLumiQC}
\alias{.plotLumiQC}
\title{Plot QC results}
\usage{
.plotLumiQC(contractid, mdata, sampledata, pdata, qctype, workdir = NULL)
}
\arguments{
\item{contractid}{Character. The name of the contract or other value}

\item{mdata}{Data Frame. A filtered data frame of M-values}

\item{sampledata}{Data Frame. The project samplesheet.Generate manually or in arraydm::readdata().}

\item{pdata}{Data Frame. A filtered data frame of detection p-values.}

\item{qctype}{Character indicating whether preQC or postQC is undertaken}

\item{workdir}{Character. Where to save the plots. (Default: working directory)}
}
\value{
Default lumi QC plots
}
\description{
A function to run lumi QC
}
