% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnorm.R
\name{.plotNorm}
\alias{.plotNorm}
\title{Plots normalised data}
\usage{
.plotNorm(plotprefix, rawdata, quantdata, swandata, sampledata, workdir)
}
\arguments{
\item{plotprefix}{The name of the contract}

\item{rawdata}{A data frame of M-values}

\item{quantdata}{Quantile normalised array.}

\item{swandata}{SWAN normalised array.}

\item{sampledata}{A data frame with the samplesheet.Generate manually or in arraydm::readdata()}

\item{workdir}{Where to save the plots. (Default: working directory)}
}
\value{
Default SWAN Normalised array and data distribution
}
\description{
Plots normalised data
}
